Split login and register into separate pages
- Created new Register.svelte component with dedicated registration form
- Refactored Login.svelte to only handle sign-in functionality
- Implemented navigation between Login and Register pages via event dispatching
- Updated App.svelte routing to support separate login/register pages
- Updated e2e tests to work with new separated authentication pages

This improves UX by providing clearer separation between sign-in and sign-up flows.
